// client/src/ulys/TokenType.ts
enum TokenType {
  CONCEPT = 'CONCEPT',

  COMMA = 'COMMA',
  QUOTE = 'QUOTE',
  STRING = 'STRING',
  SLASH = 'SLASH',
  AT_REFERENCE = 'AT_REFERENCE', // @123

  NEWLINE = 'NEWLINE',
  BACKSLASH = 'BACKSLASH',
  DOUBLE_BACKSLASH = 'DOUBLE_BACKSLASH',

  LEFT_BRACKET = 'LEFT_BRACKET',
  RIGHT_BRACKET = 'RIGHT_BRACKET',

  // Arrows
  LEFT_ARROW = 'LEFT_ARROW',
  RIGHT_ARROW = 'RIGHT_ARROW',
  UP_ARROW = 'UP_ARROW',
  DOWN_ARROW = 'DOWN_ARROW',

  LEFT_RIGHT_ARROW = 'LEFT_RIGHT_ARROW',
  UP_DOWN_ARROW = 'UP_DOWN_ARROW',

  LEFT_UP_ARROW = 'LEFT_UP_ARROW',
  LEFT_DOWN_ARROW = 'LEFT_DOWN_ARROW',
  RIGHT_UP_ARROW = 'RIGHT_UP_ARROW',
  RIGHT_DOWN_ARROW = 'RIGHT_DOWN_ARROW',

  BOF = 'BOF',
  EOF = 'EOF'
}

export default TokenType;